StoriNewsletter/requirements.txt
asgiref==3.7.2
boto3==1.34.29
botocore==1.34.29
cffi==1.16.0
cryptography==42.0.1
Deprecated==1.2.14
Django==5.0.1
django-cors-headers==4.3.1
django-postgres-extensions==0.9.3
django-ses==3.5.2
djangorestframework==3.14.0
ecdsa==0.18.0
jmespath==1.0.1
psycopg==3.1.17
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
wrapt==1.16.0
